requirements.txt
Flask
azure-identity
azure-storage-blob
Pillow
python-dotenv